requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Validation and Settings
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# CV Parsing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0

# NLP and Data Extraction
spacy==3.7.2

# URL and Data Validation
validators==0.22.0
requests==2.31.0

# File Handling
aiofiles==23.2.1

# Testing
pytest==7.4.3
httpx==0.25.2
pytest-asyncio==0.21.1

# Utilities
python-dateutil==2.8.2

# Phase 2: Data Collection
beautifulsoup4==4.12.2
lxml==4.9.3
html5lib==1.1
newspaper3k==0.2.8
